# docs/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gnovium Knowledge OS API",
    "version": "1.0.1",
    "description": "Gnovium API is a local-first Knowledge Operating System API for building applications where knowledge behaves like a living system rather than a collection of disconnected documents.\n\nThis API supports:\n- Block-based content management with rich text editing\n- Structured relational knowledge modeling with typed entities and properties\n- Interactive knowledge graph with traversal, pathfinding, and queries\n- Git-inspired versioning with branches, snapshots, and visual diffs\n- AI-powered semantic search and natural language question answering\n- Workspace governance with health scoring and duplicate detection\n- Cross-device synchronization with conflict resolution\n- File management with presigned URL uploads (cloud mode)\n\nDeployment modes:\n- **Local Mode** (default): SQLite storage, offline-first, local AI via Inference Runtime\n- **Cloud Mode**: PostgreSQL, S3 storage, managed infrastructure",
    "contact": {
      "name": "Gaurav Kaloliya — Founder & Creator",
      "url": "https://www.linkedin.com/in/gaurav-kaloliya-b44569417",
      "email": "gaurav@gnovium.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/GauravKaloliya/gnovium/blob/main/LICENSE"
    },
    "termsOfService": "https://gnovium.com/terms"
  },
  "servers": [
    {
      "url": "https://api.gnovium.com/api/v1",
...
    },
    {
...
    },
    {
...
    }
  ],
  "paths": {
...
    "/relations/": {
      "get": {
        "tags": [
          "Relations"
        ],
        "summary": "List relations",
        "description": "Returns graph relations for a workspace.",
        "operationId": "relationsList",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Workspace scope for the operation."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Page number. Defaults to 1."
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Page size from 1 to 100. Defaults to 25."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Relations"
        ],
        "summary": "Create relation",
        "description": "Creates a typed edge between two entities. Relation types are free-form strings: references, depends_on, part_of, related_to, implements, extends, or custom.",
        "operationId": "relationsCreate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained via OAuth2 or direct login. Format: Bearer <token>"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for server-to-server authentication. Generate from workspace settings."
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.gnovium.com/oauth/authorize",
            "tokenUrl": "https://api.gnovium.com/oauth/token",
            "scopes": {
              "workspace:read": "Read workspace metadata and content",
              "workspace:write": "Create, update, and delete workspace content",
              "workspace:admin": "Manage workspace settings, members, and billing",
              "user:read": "Read user profile information",
              "user:write": "Update user profile settings",
              "offline_access": "Receive refresh tokens for long-lived access"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      },
      "DataEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
    }
  },
...
}